Use PyList_SetSlice API function instead of PySequence_DelSlice.
Also report a guess on the actual number of items purged in the case
that clearing the list for some reason didn't succeed.

It appears that PyList_SetSlice does the same thing in our case as
PySequence_DelSlice, but faster. The former uses a more direct
implementation for the list object while the latter uses the PyMapping
API with real Slice object.