Use a test-and-TAS protocol to guard against contested purging.
The purging entrance protocol now uses C11 atomic feature if the feature
is enabled by the compiler. The idea is the same: in the event of
incomplete GIL protection because of callback causing GIL release,
prevent two threads from racing to purge. A purger first takes the hint
and will not use the TAS instruction if the purge *may* be busy.