Implement safe finalizer support.
This gives the callback, if any, the last chance to run safely on a
non-empty purge queue at the time of finalization. The non-empty purge
queue happens very infrequently, but remains a possibility in concurrent
code (or induced artificially). If this happens it may defer garbage
collection (by moving the cyclic object to a higher generation) and
there's a risk of "resurrecting" the object, but is otherwise safe.

See also:
- PEP 442
- bpo-33909 https://bugs.python.org/issue33909