BUG: Some exceptions should not be swallowed by the purge function.
The following exceptions should demand the attention of the interpreter
and should not be suppressed if raised during callback action. These
are:

- RecursionError
- SystemError
- MemoryError
- SystemExit

The methods that calls for a purge just before returning now check for
these conditions and return an appropriate "error" value (e.g. -1 or
NULL), and rolls back the refcount if appropriate.