Implement "TinySet" for safe type lookup.
The "TinySet", a minimal implementation of "frozenset"-like data
structure entirely in a directly-included C file, replaces the
complicated condition in an "if" statement. This leads to less branchy
and less nesty code that disproportionally affect one single function.

The assumption is that the default "safe types" do not change their
pointer value throughout the lifetime. A perfect hash function for the
fixed collection of pointers is determined at module-init-time for some
computation cost, but membership-testing with the set is very efficient.

Currently, there's no direct way to free the memory allocated at
module-init-time for the TinySet of safe types, but this can be changed
by converting to multi-stage initialization, to be followed.